fix(cmake): resolve strict MinGW compiler errors and CMake configurations
- Refactored CMakeLists.txt to consistently strip MSVC /W[0-4] warning flags globally rather than via local scopes, eliminating native D9025 overrides.
- Fixed integer sign conversion (-Wsign-conversion) across http_winhttp.c and http_wininet.c during size_t-to-pointer arithmetic for wide-character conversion, and corrected WinHttpSetTimeouts DWORD/int type mismatches.
- Explicitly cast unused return values to (void) to resolve -Wunused-result enforcement on cleanup routines.
- Hardened mock definitions in process.c with #ifndef _WIN32 guards, resolving -Wunused-function failures under strict MinGW cross-compilation.

+++ src/http_wininet.c
@@ -142,28 +142,28 @@ static int headers_to_wide_block(const struct HttpHeaders *headers,
   for (i = 0; i < headers->count; ++i) {
     size_t written = 0;
     /* Key */
-    if (ascii_to_wide(headers->headers[i].key, p, total_wchars - (p - buf),
-                      &written) != 0) {
+    if (ascii_to_wide(headers->headers[i].key, p,
+                      total_wchars - (size_t)(p - buf), &written) != 0) {
       free(buf);
       return C_ABSTRACT_HTTP_ERR_IO;
     }
     p += written;
 #if defined(_MSC_VER) && !defined(__INTEL_COMPILER)
-    wcscpy_s(p, total_wchars - (p - buf), L": ");
+    wcscpy_s(p, total_wchars - (size_t)(p - buf), L": ");
 #else
     wcscpy(p, L": ");
 #endif
     p += 2;
 
     /* Value */
-    if (ascii_to_wide(headers->headers[i].value, p, total_wchars - (p - buf),
-                      &written) != 0) {
+    if (ascii_to_wide(headers->headers[i].value, p,
+                      total_wchars - (size_t)(p - buf), &written) != 0) {
       free(buf);
       return C_ABSTRACT_HTTP_ERR_IO;
     }
     p += written;
 #if defined(_MSC_VER) && !defined(__INTEL_COMPILER)
-    wcscpy_s(p, total_wchars - (p - buf), L"\r\n");
+    wcscpy_s(p, total_wchars - (size_t)(p - buf), L"\r\n");
 #else
     wcscpy(p, L"\r\n");
 #endif
